requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
